+++ java2c/src/test/java/testclasses/fields/SetFieldShort.java
@@ -4,19 +4,19 @@ import it.se.obfuscator.Obfuscate;
 
 public class SetFieldShort
 {
-    private int fieldS;
+    private short fieldS;
 
     public SetFieldShort()
     {
         this.fieldS = 3000;
-        this.setField(-3000);
+        this.setField((short)-3000);
     }
 
-    public int getField()
+    public short getField()
     {
         return this.fieldS;
     }
 
     @Obfuscate
-    public void setField(int value){this.fieldS = value;}
+    public void setField(short value){this.fieldS = value;}
 }
