+++ java2c/src/test/java/testclasses/fields/SetStaticShort.java
@@ -4,21 +4,21 @@ import it.se.obfuscator.Obfuscate;
 
 public class SetStaticShort
 {
-    private static int fieldS;
+    private static short fieldS;
 
     public SetStaticShort()
     {
         SetStaticShort.fieldS = 3000;
-        this.setStatic(-3000);
+        this.setStatic((short)-3000);
     }
 
-    public int getStatic()
+    public short getStatic()
     {
         return SetStaticShort.fieldS;
     }
 
     @Obfuscate
-    public void setStatic(int value)
+    public void setStatic(short value)
     {
         SetStaticShort.fieldS = value;
     }
